Migrate 'not-has-class' test
